+++ CHANGES.rst
@@ -93,6 +93,8 @@ Unreleased
 -   The default status code for ``append_slash_redirect`` is 308 instead
     of 301. This preserves the request body, and matches a previous
     change to ``strict_slashes`` in routing. :issue:`2351`
+-   Fix ``ValueError: I/O operation on closed file.`` with the test
+    client when following more than one redirect. :issue:`2353`
 
 
 Version 2.0.3