+++ CHANGES.rst
@@ -47,6 +47,8 @@ Unreleased
     -   Remove ``format_string``, use :class:`string.Template` instead.
     -   Remove ``escape`` and ``unescape``. Use MarkupSafe instead.
 
+-   The ``multiple`` parameter of ``parse_options_header`` is
+    deprecated. :pr:`2357`
 -   Rely on :pep:`538` and :pep:`540` to handle decoding file names
     with the correct filesystem encoding. The ``filesystem`` module is
     removed. :issue:`1760`